Add files via upload
Updated to include the 99/1 winsorized data for the quiet standing, dual comms, and serial 7's factorial analysis. Also updated to include a variable for a chi-square analysis where 1 = increases 95% ellipses area compared to control and 0= decreases 95% ellipses area compared to control. Nylon belt is NOT included in this data set.